Split PrivateKeyFile into different implementations.
